refactor: move NetworkUtils to a new package
- Changed the package of NetworkUtils from util to util.network.
- Updated import statements in multiple services to reflect the new package structure.
