+++ app/src/main/java/fr/iut_rodez/pathpilot_android_client/home/clients/ClientService.java
@@ -1,6 +1,6 @@
 package fr.iut_rodez.pathpilot_android_client.home.clients;
 
-import static fr.iut_rodez.pathpilot_android_client.util.NetworkUtils.getRequestQueue;
+import static fr.iut_rodez.pathpilot_android_client.util.network.NetworkUtils.getRequestQueue;
 import static fr.iut_rodez.pathpilot_android_client.util.VolleyErrorHandler.handleError;
 
 import android.app.ProgressDialog;
