+++ app/src/main/java/fr/iut_rodez/pathpilot_android_client/util/VolleyErrorHandler.java
@@ -13,6 +13,7 @@ import com.android.volley.VolleyError;
 import java.nio.charset.StandardCharsets;
 
 import fr.iut_rodez.pathpilot_android_client.R;
+import fr.iut_rodez.pathpilot_android_client.util.network.NetworkUtils;
 
 /**
  * Utility class to handle Volley errors and provide user-friendly error messages
